modules/location-producer/requirements.txt
Flask==1.1.1
flask-accepts==0.10.0
flask-cors==3.0.8
Flask-RESTful==0.3.7
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
SQLAlchemy==1.3.19
psycopg2-binary==2.8.5
flask-restx==0.2.0
Flask==1.1.1
werkzeug==0.16.1
grpcio==1.41.0
protobuf>=3.5.0.post1, < 4.0dev
kafka-python==2.0.2
GeoAlchemy2==0.8.4
shapely==1.7.1
